Move strptime polyfill to string_util
