+++ src/string_util.h
@@ -54,6 +54,12 @@ template <class T> std::string str(T value);
 //-------------------------------------------------------------------
 // Utilities
 //-------------------------------------------------------------------
+#ifdef __MSYS__
+//
+// Polyfill for strptime function.
+//
+char* strptime(const char* s, const char* f, struct tm* tm);
+#endif
 //
 // Convert string to off_t.  Returns false on bad input.
 // Replacement for C++11 std::stoll.
