--- src/strptime.h
@@ -1,50 +1,0 @@
-
-/*
- * Polyfill for strptime function
- *
- * This source code is from musl library, located at https://git.musl-libc.org/cgit/musl/tree/src/time/strptime.c .
- */
-
-/*
- * musl as a whole is licensed under the following standard MIT license:
- * 
- * ----------------------------------------------------------------------
- * Copyright © 2005-2020 Rich Felker, et al.
- * 
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
- * 
- * The above copyright notice and this permission notice shall be
- * included in all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
- * CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
- * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- * ----------------------------------------------------------------------
- */
-
-#ifndef S3FS_STRPTIME_H_
-#define S3FS_STRPTIME_H_
-
-#include <stdlib.h>
-#include <langinfo.h>
-#include <time.h>
-#include <ctype.h>
-#include <stddef.h>
-#include <string.h>
-#include <strings.h>
-
-#include "common.h"
-
-char* strptime(const char* s, const char* f, struct tm* tm);
-
-#endif
